+++ src/core/const.js
@@ -36,12 +36,45 @@ class Const extends _Size { // implicit extend Numeric
       ? this.num.toString() + '.0'
       : this.num.toString();
   }
+  // Actually this is not bind but just checking after loading all components
+  // It checks lower<=num<=upper, 0<num if scale=='log', 0<num<1 if scale=='logit'
+  bind(namespace){
+    super.bind(namespace);
+    let logger = this.namespace.container.logger;
+
+    // should be: lower <= num
+    if (this.lower !== undefined && this.lower > this.num) {
+      let msg = `Constant "${this.index}" is outside of borders: ${this.num}(num) < ${this.lower}(lower)`;
+      logger.error(msg, {type: 'BindingError', space: this.space});
+    }
+    // should be: num <= upper
+    if (this.upper !== undefined && this.upper < this.num) {
+      let msg = `Constant "${this.index}" is outside of borders: ${this.num}(num) > ${this.upper}(upper)`;
+      logger.error(msg, {type: 'BindingError', space: this.space});
+    }
+    // for scale=='log' should be: num > 0
+    if ((this.scale === 'log' || this.scale === 'logit') && this.num <= 0) {
+      let msg = `Constant "${this.index}" ${this.num}(num) is not positive that is not allowed for "log" and "logit" scale`;
+      logger.error(msg, {type: 'BindingError', space: this.space});
+    }
+    // for scale=='logit' should be: num < 0
+    if (this.scale === 'logit' && this.num >= 1) {
+      let msg = `Constant "${this.index}" ${this.num}(num) is not less than 1 that is not allowed for "logit" scale`;
+      logger.error(msg, {type: 'BindingError', space: this.space});
+    }
+  }
   clone(){
     let clonedComponent = super.clone();
     if (typeof this.num !== 'undefined')
       clonedComponent.num = this.num;
     if (typeof this.free !== 'undefined')
       clonedComponent.free = this.free;
+    if (typeof this.scale !== 'undefined')
+      clonedComponent.scale = this.scale;
+    if (typeof this.lower !== 'undefined')
+      clonedComponent.lower = this.lower;
+    if (typeof this.upper !== 'undefined')
+      clonedComponent.upper = this.upper;
       
     return clonedComponent;
   }